Save query counts.
Currently, the results of multiple queries fed to this node are not distinguishable. A key `queryCounts` has been added to `msg.payload` for subsequent nodes to know count of rows returned by each query. The sequence of counts is same as that of queries that is input to this node.